+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.1.0] - 2017-08-09
+### changed
+- Changes method _get_default_code in model of products for assign the same SKU to variants of the same product template. Also changes the form view of products (templates) for always show the SKU (internal reference).
+- Adds a system parameter for the number of digits to be used in the secuence of SKU of the products.
+
+
+
 ## [1.0.0] - 2017-07-25
 ### added
 - Extends addon product to automate the generation of SKU code with reference of supplier in field default_code of products.